Allow building without inlining in debug mode under MSVC
Because there's a reason it's normally disabled by default.

Inlining in debug mode is enabled by default for the MSVC build of HHVM because it significantly reduces the size of the static libraries, however I hit an issue where I needed to disable it globally, so, rather than a temporary hack, I added a proper config option for it.