Mark the 'MemExceededFlag' as not sticky
This diff changes the surprise flag that denotes that the current
request has run out of memory as not 'sticky', meaning that it will
now be cleared when surprise flags are checked and cleared. It should
_not_ be sticky because this will cause all subsequent calls to OOM,
even after the stack has been unwound and everything cleared (for
example, this causes shutdown handlers to OOM deterministically even
when there is ample memory).

The previous behavior was likely an unintended consequence of commit
40f9d9547b3e13767575534686e5e1b061d2d029.