Fix places where MSVC warns about truncation of literals in a cast
Summary: I've ignored the couple of places where MSVC complained about this so far, because there is no difference, at least, not in these cases, in the resulting value if done on a signed or unsigned value. However, c8c6fb91 added a spot in `hhbc-codec.h` that triggered this warning, and, because it is in a template, the warning ends up showing up everywhere, and that finally prompted me to create the PR to eliminate them.
Closes https://github.com/facebook/hhvm/pull/6223

Reviewed By: @JoelMarcey

Differential Revision: D2432694