Fix MixedArray::InitSmall on platforms not using the inline ASM
Summary: This was a massive pain to track down.
The issue here is that `hashTab()` is calculated based on the scale value of the array, but that doesn't get set until farther down. The solution to this is simple, use the `mixed*` helper funcs, which are defined for this exact purpose.
Closes https://github.com/facebook/hhvm/pull/6204

Reviewed By: @JoelMarcey

Differential Revision: D2419073