Add `hhvm.expect[f]` as accepted file suffix for test runner
Summary: Since we are allowing `typechecker.expect[f]` for running the typechecker
with the test runner, this provides an explicit `hhvm.expect[f]` suffix for
HHVM-based tests. This new suffix is a synonym for the original `.expect[f]`
files.

The impetus of this is for the examples in the Hack user documentation which
we want to be able to be tested using both HHVM and the typechecker. For users
it would be good to distinguish between the two files.

Reviewed By: @fredemmott

Differential Revision: D2415358