Support `.php.type-errors` as a test runner file suffix
Summary: Allow the test runner to support `.php.type-errors` as a file suffix. This type of Hack
or PHP source file is only supported in `--typechecker` mode. The use case is our user
documentation where we want to show a few examples with type errors, but we want `hh_client`
to still run clean when run over the entire repo. You can use the test runner to check that
the errors we claim is output is actually output.

Depends on D2401016

Reviewed By: @fredemmott

Differential Revision: D2416180