Port ext_fileinfo to MSVC
This pokes the condition around a couple of includes and defines, adds a set of `#undef`'s to config.h, and deals with the fact that `time_t` is a 64-bit value and `ts.tv_sec` is only a `long` value.

This extension actually requires quite a bit more poking of the includes, which will come as part of #6148 (D45663), due to their dependence on the upstream folly change.