Fix missing decrefs in inc/dec operations.
Summary: In the interpreter: IncDecN, IncDecG, getS, IssetS, and EmptyS
failed to decref the name operand. IncDecS and IncDecL had it
right; handle the decref in SpropState for the *S opcodes, and
explicitly for the other IncDec opcodes.

In ObjectData::incDecProp: All the code paths that invoked a
custom getter prior to inc/dec failed to decref the intermediate
value returned by the getter. These seem to have been broken ever
since the ObjectData/Instance merger a long time ago.

Reviewed By: @alexmalyshev

Differential Revision: D2433654