Move MemExceededFlag out of ResourceFlags
This diff moves the MemExceededFlag out of ResourceFlags but keeps it
in StickyFlags. This keeps the flag from being cleared inadvertently
as intended in 1a0b6ab85b6ece5c42ccdbc86a05d461b1c1ec0b but it ensures
that the flag wont be preserved across calls to handle_exception_helper,
which preserves all surprise flags in ResourceFlags. The
handle_exception_helper is the sole consumer of ResourceFlags.