UVS: Split parser into PHP5 and PHP7 versions
Summary: Due to the BC breaks in UVS, we want to have two parsers, one
for PHP5 and one for PHP7 mode.

The //vast// majority of the rules are the same between the two, so just
copy-pasting is too heavyweight. The differences are actually just in a
few productions in a few nonterminals. I originally was going to splice
in different segments in each mode, but even that turned out to be too
heavyweight. With such similarity, IMO it's easiest to read and reason
about just using special comments to distinguish which bits need to go
in which parser. This diff does that, augmenting `make-parser.sh` with
some awk magic to strip out the irrelevant bits in the other parser.

To keep this reviewable, no actual parser changes are in this diff, just
the machinery to make it happen. I have a few diffs which take advantage
of the split almost ready in my sandbox, so stand by for dependent diffs
to get put up later today if you want to see what this looks like in
action.

Part of https://github.com/facebook/hhvm/issues/5523

Reviewed By: @paulbiss

Differential Revision: D2380191