Fix interpOneCFHelpers implementation
Summary: The interpOneCFHelpers are supposed to have resumeHelper-like behavior,
but instead they often has interpHelper-like behavior because we aren't
setting the `interpFirst` argument register to 0.

This bug was introduced in the vasm conversion, when the jmp to
us.resumeHelper was accidentally changed to a jmp to rh.handle_resume---
which used to be an X64Assembler Label inconveniently also named
"resumeHelper".

Reviewed By: @ottoni

Differential Revision: D2443293