Replace alloca() in the interpreter with req::vector
Summary: This was bad bad bad bad.  In an optimized build, the iop helper function would
get inlined into the interpreter's dispatch loop, increasing the lifetime of the
alloca() to the function containing the loop.  We'd keep growing our stack with
each call of the inlined helper, eventually blowing it in some cases.

Reviewed By: @edwinsmith

Differential Revision: D2444667