Provide a typing hook for parent::__construct expressions
Summary: There was no typing hook for informing a client that a super class constructor call has been encountered. This appears to be an oversight. Since the type checking code for this case is very different from other qualifier::name(...) expressions, a new hook is introduced to keep things simple at the call site and to avoid any disruption to callers of existing hooks.

Reviewed By: @int3

Differential Revision: D2442694