Refactor CodeGenerator, pt. 1: namespacing and arith
Summary: Begin the work of splitting up codegen into multiple files and
extricating it from the Verber class idiom.

This diff delegates some of the most basic codegen helpers, like
srcLoc(), as well as the arithmetic HHIR instructions, to a new irlower
namespace (mirroring irgen), with the ultimate goal of doing likewise
with the rest of CodeGenerator.

Reviewed By: @ottoni

Differential Revision: D2427600