Make str_repeat consistent with upstream PHP
Summary: Script:

    <?php

    $var = str_repeat('A', -1);
    var_dump($var);
    $var = str_repeat('A', 9223372036854775807);
    var_dump($var);

See https://3v4l.org/qpA7N for the results.

In PHP 7, the script returns NULL / warns for the first str_repeat and fatals on the second. In PHP 5 it warns/returns null for the negative one and warns/returns the empty string for the second one. I decided to defer to the PHP 7 behavior.

I also changed the type for the function multiplier argument (otherwise we see overflow if the multiplier is too large and we treat it as a negative number).

Note that the HHVM max string size and PHP max string size appear to differ.

Reviewed By: @markw65

Differential Revision: D2437850