Teach load-elim about CheckInitMem.
Summary: It didn't know how to fold it before.  Also adds a simplifier pass at the very
end of optimize() to clean up the results of load-elim, store-elim, etc.  This
pass is only enabled when we're not in a profile translation since that's when
those passes would have actually done any work.

Reviewed By: @swtaarrs

Differential Revision: D2446237