Hide the master PHP7 option
Summary: Nothing in HHVM should depend on this option directly, use the
granular ones instead. I'm really worried this will turn into the mess
that HipHopSyntax is where you don't have any idea all the things that
it does. So hide the option so no code outside of RuntimeOption can see
it.

Reviewed By: @paulbiss

Differential Revision: D2444632