Preserve nullable field for type structures with type alias
Summary: There is a bug currently when resolving type structures containing aliases
where we do not preserve the nullable field. For example

```
type MyAlias = int;
class C {
  const type T = ?MyAlias;
}

type_structure(C::class, 'T');
```

The type structure for `C::T` will not be nullable because we take the type
structure directly from the alias. This adds a simple check that tests if the
original type structure was nullable, then the alias resolved type structure
should be nullable as well.

Reviewed By: @paulbiss

Differential Revision: D2452841