Convert ext_gettext to HNI
This converts the very simple gettext extension from `ext_zend_compat` to use HNI instead.

I originally did the port because of me failing to realize where libintl was being searched for, but this is a simple enough extension that I didn't see any reason not to submit a PR to switch it over.