Port ext_zend_compat to MSVC
This required a lot of poking of various headers.

The most notable difference from the PHP version of these is that pthread locking is used for zend compat under MSVC for HHVM, where-as PHP uses native windows locking. This was done because there are multiple APIs that are reliant on the pthread locking, and it was easier just to change that than deal with special casing those APIs for MSVC.