Rename .type-errors files when running test
Summary: We have files with the `php.type-errors` extension in the docs so
that `hh_client` will run clean over all of the examples, but that
the test runner can be used to show that type errors actually exist
in those files.

However, when running the test runner test for such a file, those files
need to be renamed back to just `.php` only since `hh_server --check`
understand that suffix, not `.type-errors`.

This diff patches the test runner to make sure that rename happens.

Depends on D2447201

Reviewed By: @fredemmott

Differential Revision: D2449480