Allow builtins to use better calling conventions
Summary: Passing Strings, Arrays, Objects and Resources
currently ends up passing a pointer to the pointer to
the FooData. Instead we can just pass the pointer.

This diff creates a "Dumb-pointer" class that wraps a pointer.
It has no copy constructor, or destructor, so it can be passed
by value - effectively equivalent to passing the pointer.

It also allows you to pass Mixed parameters in a TypedValue
rather than as a const Variant&, and to return Mixed in
a TypedValue (so in registers on x86).

Finally for output parameters, it allows the param to be passed
as a RefData*; but note that in this case, if the parameter was
not passed by reference, nullptr will be passed. A later diff
could add support for inout params.

Reviewed By: @swtaarrs

Differential Revision: D2230402