AsyncMysqlConnection::multiQuery takes both an array and a Vector
Summary: When documenting an example for `mulitQuery`, I used an `array` to pass in
the queries. It worked in HHVM. I also tried to pass a `Vector` and it worked too.

The HHVM HNI definition specifies an array.

```
  <<__HipHopSpecific, __Native>>
  function multiQuery(array $queries,
                      int $timeout_micros = -1): ExternalThreadEventWaitHandle;
```

However, when typechecking my code, I get an error:

```
multiQuery.php:29:38,45: Invalid argument (Typing[4110])
  /tmp/hh_server/hhi_185c7d40/hhi/stdlib/builtins_async_mysql.hhi:86:30,35: This is an object of type Vector
  multiQuery.php:22:14,18: It is incompatible with an array (used like a vector)
```

The `hhi` file had it taking a `Vector`. Make it take a `Traversable`.

Reviewed By: @​@jwatzman

Differential Revision: D2454151