Ensure we have a standalone serial test bucket and make specifying serial tests more generic
Summary: We didn't have a standalone serial test bucket before.

When running `test/run test/slow/ext_mysqli/ test/slow/ext_memcache test/slow/ext_memcached/`

See the buckets in our current system; look at bucket 4. It contains the serial tests, but it also had a parallel test in it (`return_types`).

P20039027

The new way ensures that there is a bucket for serial tests only. See bucket 31. Only the 8 tests that are supposed to be serial are there.

P20039029

This diff also adds a `.serial` file for those tests that will be put in the serial bucket.
Now we don't have to hard code those tests in the code and we can easily add
more serial tests with the `touch` of a file.

Use case: I want to run some of our documentation tests in serial, particularly
the async mysql ones.

Reviewed By: @jwatzman

Differential Revision: D2469773