flushAll returns mc_res_ok, not mc_res_stored (isStored())
Summary: We are throwing an exception even when `flushAll` succeeds:

```
Fatal error: Uncaught exception 'MCRouterException' with message 'flushall failed with result mc_res_ok' in /data/users/joelm/user-documentation/guides/hack/99-api-examples/mcrouter-examples/mcrouter-examples/flushAll.php:53
Stack trace:
```

This is because we were checking for the wrong reply value. This changes that so we check for the right one.

Reviewed By: @pavlo-fb

Differential Revision: D2474930