Get rid of ServerWarmupThrottleRequestCount
Summary: For the first a few threads, we have the same number of worker threads
as the CPU count.  The purpose of this arrangement is to ensure that the
thread that is jitting can receive some CPU.

This diff tries to achieve the same thing by bumping the priority of the
thread that holds the write lease for jitting.  This might even make
warm-up faster.  At least it should give us some more capacity in the
first minute after server starts up, although we will be mostly
interpreting.

This is the cherry-pick for D2487836.

Reviewed By: @binliu19

Differential Revision: D2511539