Infer the type of array() initialized expression based on their further usage
Summary: Currently, we just give up and never track anything about array() expressions. After this diff, [] turns it into vector-like array and [...] makes it a hashtable-like array, after which the types of it's values/keys can be successfully tracked. This has few "false positives" (or rather, "type-checker limitations", for example when someone uses:

  $a = array();
  for($i = 0; $ i < n; $i++) {
    $a[$i] = ...
  }

to create a vector we will still think that it's a hashtable, but overall seems much more beneficial than annoying.

Reviewed By: @dlreeves

Differential Revision: D2446230