Support Windows and a non-pcntl mode in the tester
This has two distinct parts to it.
The first is adding support for a non-pcntl serial mode. This mode outputs the results of each test directly.
The second are a few changes for Windows, dealing with the directory separator character, and the timeout script not working under msysgit's bash shell.