Allow type structure to work with type aliases
Summary: Currently `type_structure()` only works for retrieving the TypeStructure for a class's type constant. We also create type constants for type aliases which are accessible via `ReflectionTypeAlias`. There are cases where we will want to avoid creating the reflection object and access the the type structure directly. This overloads the `type_structure()` function so that if it is provided with a single argument it will fetch the TypeStructure assoicated with a type alias.

Reviewed By: @paulbiss

Differential Revision: D2450585