Add 'typename' as an AliasEntry in the parser
Summary: We want a way to reference the name of a type alias in Hack. Since
TypeAlias::class already works we will utilize the string returned by that
expression. The type given to this will be `typename` since it doesn't reference
an actual name of a class.

This is done via the same mechanism we use for the `classname` type by adding an AliasEntry in the parser.

Reviewed By: @jwatzman

Differential Revision: D2473483