Properly resolve generics for type alias' type structures
Summary: Currently type structures do not properly represent aliases with generics. For instance:

```
type MyMap<T> = Map<T, T>;

type StrMyMap = MyMap<string>;
```

When we resolve the type structure for `StrMyMap` it will be `Map<T, T>` instead of `Map<string, string>` which is what we expect. This diff fixes this issue by substituting in the proper type for the generic `T`. When we resolve a type structure we need to pass in a mapping from the name of generic to the type it maps to. This is created whenever we attempt to resolve an alias. We split the type_vars field stored by the delimiter `,` which will give us an ordered list of type parameters. We then look at the generic_types field of the type structure we are resolving and create the mapping from generic name to type.

Reviewed By: @swtaarrs

Differential Revision: D2459004