Define 'zero' vasm instruction to clear register
Summary: Some architectures might not use xor as an effective way of
zeroing a register. On top of that it forbirds xor from being lowered as
the Vxls::insertCopiesAt is called after lowering step. It'd be good to
have a specific vasm instruction for doing that.

This approach will give the flexibility needed for other architectures
that might want to zero differently or lower the xor vasm instructions.