Properly handle kills in store-elim
Summary: We weren't unsetting mayLoad bits for the kill sets on various
MemoryEffects specializations, causing us to miss out on a lot of the benefits
of the kill sets. The main consequence of this was that we weren't eliminating
any stores to locals/stack that were only visible to a RetCtrl. My upcoming
diff to remove FrameState::EvalStack is a regression without this fix, which is
how I discovered the issue.

I also fixed AliasAnalysis::expand() to handle AStackAny; previously it only
handled specialized subclasses.

Reviewed By: @markw65

Differential Revision: D2486299