Remove EvalStack from FrameState
Summary: EvalStack is now gone, so we no longer do any store elimination for
stack slots during gen time. We rely on the store-elim pass to clean things up
later (which turned out to be a slight performance win).

This removes a lot of code and simplifies a number of things, particularly some
new minstr stuff I have going in another diff.

Reviewed By: @alexmalyshev

Differential Revision: D2489098