Wait until dfind is ready before considering init done
Summary: Anecdotal evidence suggests that dfind can take a long time to initialize, so
the main hh_server process ends up blocking on it after startup. Recently, we
added a timeout to dfind requests because they were timing out, but I'm
guessing that many of these apparent timeouts are just the first request to
dfind taking a long time to return -- Scuba logs show that the wall_age of
these timeouts is generally on the small side, indicating that hh_server
started up recently.

This diff adds a new INIT_REALLY_END event that waits for dfind to be fully
ready. It also has dfind log its initialization time to Scuba.

Reviewed By: @dlreeves

Differential Revision: D2497055