Fix timezone transitions
If there were no transitions, the function would segfault (this can
happen with timezones like UTC), so it now avoid that.

Also makes the behaviour when passing a starting timestamp consistent
with Zend PHP. Previously the code always used the first transition, no
matter what starting time was given. Not only is this unexpected and
mostly useless, it is also inconsistent with Zend behaviour, which
uses the most recent transition from before the start time.

It also makes the code a little more robust.