Hide the ENABLE_EXTENSION_* variables for required extensions
This hides the ENABLE_EXTENSION_* variables from the CMake GUI on Windows, because you can't disable those extensions to begin with.

This was originally [#5941](https://github.com/facebook/hhvm/pull/5941), but that got shot down quickly, as it was removing the actual variables. This keeps the variables but hides them from the UI.