Add ability to precisely scan native data objects.
Summary: Add ability to precisely scan native data objects.

Add an entry for a precise GC scan function to NativeDataInfo.  If a class registered with Native::registerNativeDataInfo has a scan(IMarker&) method, it will get added to the NativeDataInfo automatically.  If the class has no scan function, it will default to conservative scanning.

Reviewed By: @edwinsmith

Differential Revision: D2357304

fb-gh-sync-id: fcb9346a69c93593212084aaefa83a040aeb3f56