Fix persistent SSL/TLS sockets
Summary: Resolves #5348. Fixes SSL/TLS sockets not being usable when a socket is
reused from persistent storage. A HHVM::Socket was being created when
SSLSocket was the proper type to create.

Please let me know if you would prefer something different in place of the dynamic cast, like a virtual method on Socket, to help determine if the thing is an SSLSocket.
Closes https://github.com/facebook/hhvm/pull/6329

Reviewed By: @fredemmott

Differential Revision: D2508173

fb-gh-sync-id: d5f7d533e0a18857418e50c91d811597f6504edf