Fix up init_end event
Summary: In {D2497055} I meant to keep the INIT_END event while adding a new
INIT_REALLY_END, but I actually deleted INIT_END -_-

This fixes it up. It also removes the INIT_END event that we used to record
after saving state (we will now consider these to be just "fresh" inits) --
because of the way the code is structured now, measuring that with a separate
event would be easier. However, I've never needed to look up save times, so
I'm not going to bother adding that for now.

Reviewed By: @dlreeves

Differential Revision: D2510048

fb-gh-sync-id: 2557c260503ecfa519bf495efe7cb320cea81755