Add type checker support for typename<T>
Summary: It is possibe to access ::class for things other than classes. For instance if
you have a type alias `TAlias`, `TAlias::class` will return the qualified name
of the alias as a string. This string is a useful type to represent in Hack,
primarily in combination with the `TypeStructure` feature.

This diff adds a new specially recognized type alias called `typename`. It is
a super type of `classname` that does not support any of the special logic such
as allowing construction with the new operator or accessing static members.
Instead it simply means this a string which is the name of some type.

This type is the result of the expression `TAlias::class`. In order to support
this I needed to add a new AST node since we do not have any where to store a
constant for type alias.

Reviewed By: @int3

Differential Revision: D2523263

fb-gh-sync-id: 6fabcb4da84f18dc278bff92f1463a198c041b13