Fix coverage metric for shapes, add coverage unit tests for AKempty -> AK[map|vec] update
Summary: Broken (by me) in {D2202779} - change made typing more sound, but coverage computation less correct.

The reason is an unfortunate action at a distance going on - we create a new field in grow_shape, but defer it's assignment to assign_simple. Between those two invocations there is quiet a lot going on, including calls to functions that fire typing hooks and record this not fully complete type. I solve this by initializing the field to the same type we initialize Vector {} element type - empty Tunresolved inside Tvar, which is considered checked (as opposed to just plain Tvar).

I noticed it when I was just about to break AKempty to AKmap/vec/shape conversions in the same way, so adding regression tests for them.

Reviewed By: @dlreeves

Differential Revision: D2514994

fb-gh-sync-id: a38afd4297d4b321555b72412997f16e44b1c877