Introduce AKshape, better local inference for arrays used like shapes
Summary: If an array is used like a shape (accessed only with static strings and class constants), we treat it like a shape, separately tracking the type of each field. The moment it's used in a way that would make it unsound, it's downgraded to a map like array.

There are few places that could be handled more gracefully and corner cases I will need to follow up on, but I feel like those all can be handled in a separate diff and this part is ready to be reviewed.

Reviewed By: @dlreeves

Differential Revision: D2511073

fb-gh-sync-id: 6ffe67ab7f86df1e7afedf9f00f11ae0138b7d10