Integration tests should have a clean repo as a starting point
Summary: The current saved state supports non-clean repos, but we never
actually use it that way in practice, so it's not a very useful thing to test.
More importantly, the new mini saved state can't be built on repos with errors,
so this change will allow me to continue using the same test suite for both.

Reviewed By: @dlreeves

Differential Revision: D2521737

fb-gh-sync-id: d6e0345348294f0b96a6e3bed253abda2643e7e0