Support PHP7's unserialize() allowed_classes option
Summary: HHVM introduced a similar piece of functionality 2.5 years ago
with somewhat different semantics.

In https://github.com/facebook/hhvm/commit/6a93844442a0f014b309e11ad296f6ef90b35ffc
the whilelist argument is only validated if an INI setting is enabled,
and even then might only warn of the attempted deserialization,
but not actually block it.

This diff introduces PHP7's "always on" filtering described in:
https://wiki.php.net/rfc/secure_unserialize
(note: scroll down to see how the final version was implemented).

When performing the whitelist check, we first check to see if the
PHP7 style whitelisting is used (by the presence of an
"allowed_classes" key.
If this option is present, it takes precedence and behaves like PHP7.

If this option is not present, we then check the RuntimeOption:
UnserializationWhitelistCheck, and warn/error/continue as appropriate.

Closes https://github.com/facebook/hhvm/issues/6052

Reviewed By: @jwatzman

Differential Revision: D2510411

fb-gh-sync-id: c0d30d144aec2ab1493dd05a3f6a94dce24b50d1