Adding frame pointers to fb_setProfile
Summary: fb_setProfile does not provide an easy way of matching exits to enters. The only way to do this right now is based on function name and creating your own stack. Adding the frame pointer as part of the frame data array should make it easier to manage the context map on the PHP side.

Another difficulty in fb_setProfile is for profilers to do exclusive calculations. It was hard tracking current frame without having frame_ptr, but tracking parent frame required calling debug_backtrace or relying on execution order; hoping the previously entered frame was the parent of the current entered frame (which is a wrong thing to do since async messes up execution order).

Reviewed By: @markw65

Differential Revision: D2516526

fb-gh-sync-id: 4d74591486cbfaccf78cf443575dbd00433d1937