Hide the ENABLE_EXTENSION_* variables for required extensions
Summary: This hides the ENABLE_EXTENSION_* variables from the CMake GUI on Windows, because you can't disable those extensions to begin with.

This was originally [#5941](https://github.com/facebook/hhvm/pull/5941), but that got shot down quickly, as it was removing the actual variables. This keeps the variables but hides them from the UI.
Closes https://github.com/facebook/hhvm/pull/6345

Reviewed By: @jwatzman

Differential Revision: D2515551

fb-gh-sync-id: a9acd4f92ab32d07765de9c7d77f9dc54d97d96f