Remove old style constructor in finfo
Summary: In deprecating support for old style constructors in D2515710, it was noticed
that one of our extensions actually uses one -- `finfo`. This diff removes
that old-style constructor and replaces it with `__construct`.

Doing this will allow us to remove the "workaround this problem `isBuiltin` check"
in D2515710.

One of the currently good zend tests used it, actually, so it had to be moved
to bad.

Reviewed By: @jwatzman

Differential Revision: D2521039

fb-gh-sync-id: f171d34aa35effc59cffb6e200ed32ec2667dd22