Lightweight debug_backtrace hash implementation
Summary: Calling debug_backtrace is very expensive.
Especially when profiling, it is called a lot.
In some extreme cases, we need to aggregate metrics by the stack.
If we want to aggregate metrics by the stack an extremely expensive "keying" of the stack takes place, creating a huge key (maybe hashing it in PHP) and then doing it again when looking it up the map.

This implementation should be a lightweight function to get a stack hash that could be used as a key in aggregations. The actual debug_backtrace should only be called once per hash (by the PHP dev).

Did a micro benchmark by calling these functions 59700 times through 4 functions in the stack.
Average time for debug_backtrace 2050ms (allocates 1883MB~)
Average time for hphp_debug_backtrace_hash (UNTESTED ESTIMATION) ~44ms when all files of functions are different (allocates 1MB)
Average time for hphp_debug_backtrace_hash 22ms when all files of functions are same (allocates 1MB)
micro benchmark is in: ~duv/temp/phptests/teststack.php

Reviewed By: @markw65

Differential Revision: D2515885

fb-gh-sync-id: a6a1a445d4ec544f470284e4f3d12a4886e3c12f