Track and use known callees in HHIR
Summary: Instead of just using the known-callee information coming from the RegionDesc, this diff extends the fpiStack in FrameState to also keep track of the callees when they're written to the ActRec by SpillFrame instructions.

I'm also keeping some extra asserts and debugging that I added why hunting for a bug I was hitting with this diff (fixed by D2512136).

Reviewed By: @paulbiss

Differential Revision: D2382189

fb-gh-sync-id: c1d3b5ce8a8ebf6558eb01d35a2a06310c6c1861