Bypass warnings/notices earlier
Summary: We originally wrote this code to bypass warnings and notices as early
as possible; snprintf is actually quite expensive, and for code that
raises a lot of notices, but has them disabled (or has a large
NoticeFrequency) it can have a significant perf impact.

This was all refactored a year ago so that all the checks are
centralized in raise_message. In addition to losing the early exit, it
also broke raise_warning_unsampled, because it called into
raise_message which promptly sampled it.

Reviewed By: @alexmalyshev

Differential Revision: D2526899

fb-gh-sync-id: af05e225a24e9de792edeaada314243aacbea4ab