Move ext_closure into its own runtime/ext directory
Summary: This is mostly a straight code move with no functional changes.
The only meaningful change is to remove the CMake hard-coding
of closure.idl.json.

This is a book-keeping move in prep for conversion to HNI
so that the actual migration will be easier to review as
it won't be muddled up with this bit of nothing.

Reviewed By: @JoelMarcey

Differential Revision: D2544716

fb-gh-sync-id: 872a34b926050e1dfc5082265859c98b75d19ada