Better oom detection
Summary: We reset the oom state before running shutdown functions, and again
before running postsend functions. But because we only checked for oom
when peakUsage increases, we wouldn't necessarily detect a second or
third oom.

The test I updated was setup to use the exact same amount of memory in
a shutdown handler, as it had used in the main part of the request
(20M vs a 5M limit), and generally, the second oom was detected. But
if the machine was particularly busy (I didn't figure out exactly how
that affects things) the second oom wouldn't be noticed because
stats.usage would be a few k less than stats.peakUsage.

After modifying the test so that it allocated 20M for the first oom,
and only 15M for the second, it reliably failed to oom the second time.

I moved the check for usage > maxBytes outside of the usage > peakUsage
check. Hopefully that won't affect performance.

Reviewed By: @paulbiss

Differential Revision: D2526776

fb-gh-sync-id: 9d9b51c180e70ca525f9ef994ff4be6ddfa6e10a