Don't use IRInstruction::become() in IRUnit::expandJmp()
Summary: expandJmp() used to indirectly use become(), through its call to
replace(). become() removes and re-adds any edges on the old and new
instructions, which can change the order of predececessor lists when the
old/new instructions are control flow. This causes problems in store-elim,
where it keeps a vector of stores in the same order as the predecessors they
come from while calling insertPhis() (in resolve_flat()).

Ideally we should change insertPhis() to take a map from pred -> value rather
than a vector of values that's assumed to be in the same order as the
predecessor list, but that will be a bigger change and come in a later diff.

Reviewed By: @alexmalyshev

Differential Revision: D2544582

fb-gh-sync-id: 744f64b46c710ffc9b415595e5fbd34229cb838c