Implement Closure::call
Summary: This implements the new `Closure::call` function in PHP 7.

`Closure::call` is essentially `Closure::bindTo` and the call to that bound
closure all in one. https://wiki.php.net/rfc/closure_apply

Also, worked around an inconsistency where `Closure:BindTo`  raises a warning and continues on if we are trying to bind an instance to a static closure. However, with `Closure::Call` we are supposed to just return null .

Fixes #6043

Reviewed By: @jwatzman

Differential Revision: D2520934

fb-gh-sync-id: f702796f104b4e0d09551326eec84db5f3ce95af