Use AsyncRetCtrl for await in resumable contexts
Summary: This has the same effect as a RetCtrl here.  RetCtrl does a vret, which
pushes the frame's m_savedRip (always set to callToExit, in a resumed
async function) and then returns to it.  callToExit then pops an item
off the native stack and jumps to enterTCExit.

We enter resumed contexts via the enterTCHelper$callTC path, so the top
of the native stack is already enterTCExit---and AwaitRetCtrl will
execute a leavetc, which simply returns to the address.

Overall, the behavior is preserved, but we do one fewer pop and avoid a
jump into a cold stub when awaiting in resumed contexts.  (Also, the ABI
gets a little cleaner.)

Reviewed By: jano

Differential Revision: D2473815

fb-gh-sync-id: 27a76bb9d486f68c073a8c5f00d7fefc72a226eb