Fix bug in ptr_subtype()
Summary: It used to think Unk was a subtype of Ref, because remove_ref(Ref) ==
Unk. Special case Unk before we get to the remove_ref() call, like other users
of remove_ref() do.

This was causing us to think that PtrToGen - PtrToRefGen was Bottom, messing up
alias analysis and causing load/store-elim to misbehave.

Reviewed By: alexmalyshev

Differential Revision: D2554135

fb-gh-sync-id: 0637ff0ae17002dbac260bd67b5718f586098d61