Fix segfault in strtr using the Wu-Manber algorithm
We would not take the source string length into consideration when calculating
the smallest string. That would result in an unsigned integer overflow
calculating the lastpos in WuManberReplacement::translate. That would later
result in a segfault trying to access a char of a string beyond its size.

Fixes #6384

Test Plan: new test to confirm fixing the segfault. Old slow and zend tests
to confirm I didn't break anything.