Port ext_zend_compat to MSVC
Summary: This required a lot of poking of various headers.

The most notable difference from the PHP version of these is that pthread locking is used for zend compat under MSVC for HHVM, where-as PHP uses native windows locking. This was done because there are multiple APIs that are reliant on the pthread locking, and it was easier just to change that than deal with special casing those APIs for MSVC.
Closes https://github.com/facebook/hhvm/pull/6285

Reviewed By: jwatzman

Differential Revision: D2464491

fb-gh-sync-id: a3a63940d4982324e40e6bc59374ba7264a799b9