Replace HHIR's relaxGuards pass with a region-level guard-relaxation pass
Summary: This diff adds a new optimizeGuardsEagerly() pass and applies it to
regions formed by the tracelet region selector.  The guards are
relaxed eagerly, without profiling information, in a similar way to
what the HHIR relaxGuards() pass was doing.  This new pass enables us
to retire the relaxGuards() pass, which had a few negative points:
(1) it was more complicated than the new pass;
(2) it was the last remaining use of FrameState outside of irgen;
(3) it was a bit fragile and had to be applied right after irgen;
(4) it changed HHIR after it was generated, so that types couldn't be
fully trusted before relaxGuards() ran, and thus was error-prone.

The type constraints used by the new pass are still obtained through
the same constrainValue & friends mechanism in IRBuilder.  We may want
to replace that with a bytecode-level type-constraining approach,
similar to what the original tracelet guard relaxation and the guard
constriction prototype used.  But that's orthogonal to this diff and
so we can experiment with that later.

This diff just turns off HHIRRelaxGuards for now; I'll kill the old
code in a separate diff for review purposes.

Reviewed By: swtaarrs

Differential Revision: D2542892

fb-gh-sync-id: 4f934e3b1fcc6b96b4b10935d52027acb30a1c09