Tweak inlining heuristics
Summary: gcc stops inlining when a "large" unit has increased by
"inline-unit-growth" percent. By default, thats 20, so inlining can be
quite limited in large units. One problem with this is that every
inline function that gets seen contributes to the initial size, and so
a lot of units are large just because it includes a number of headers
with big, unused, inline functions.

This turns up inline-unit-growth to 200%, and gets significant
wins. It should also stabilize behavior when eg we convert macros to
inline functions (which has often been a source of mysterious
regressions in the past).

I've also played with what constitutes a large function - but for now
leaving it at the default of 10000. Will do more perf testing after this
lands with other combinations.

Reviewed By: binliu19

Differential Revision: D2553689

fb-gh-sync-id: c182b000bdf94352a1497f8c5948b877d99794bc