Don't access frame locals during frame teardown
Summary: If we're in the process of tearing down a frame, and a local's
destructor fatals, we catch the fatal, and raise a warning. If there's
an error handler registered, we'll try to construct the context array
from the locals - some of which may already have been destroyed. Using
that array is likely to crash, and if it doesn't (or if the handler
doesn't use it) then destroying the array is likely to crash.

Reviewed By: mxw

Differential Revision: D2553329

fb-gh-sync-id: 011d53c46dcdd653f8023a30b202bf44ee4eb42d