Fix shifts by large integers for PHP7
Summary: In PHP5, `$x >> $y` relied on undefined behavior when `$y >= 64`.
This would tend to get compiled into an `shr` instruction on x86,
which treats that as `$x >> ($y % 64)`, which is probably not what you
wanted. PHP7 defines this to return `0`. (And similarly for right
shift.)

Part of https://github.com/facebook/hhvm/issues/6047

Reviewed By: markw65

Differential Revision: D2491999

fb-gh-sync-id: 3c1a0aa801a683df341937d4083a65cb5f3b1788