Fix negative shifts for PHP7
Summary: In PHP5, `$x >> $y` relied on undefined behavior when `$y < 0`.
This would tend to get compiled into an `shr` instruction on x86,
which treats `$y` as an unsigned int, which is certainly not what you
wanted. PHP7 defines this to throw an `ArithmeticError`; given that we
don't have `ArithmeticError` until
https://github.com/facebook/hhvm/issues/6012 is finished, I throw an
`InvalidOperationException` instead.

Part of https://github.com/facebook/hhvm/issues/6047

Reviewed By: markw65

Differential Revision: D2505138

fb-gh-sync-id: 0cd209d0cf892933d81684fcf28e5b9572fd4588