[hack[ Fix hh_find_lvar_refs IDE function
Summary: {D1576504} - sped it up by doing only naming
{D2054071} - moved the thing that hh_find_lvar_refs needs to happen after naming

So basically reverting D1576504 (which, according to at the comment there, will also be a 5x slowdown ¯\_(ツ)_/¯ )

Note: some other (or the same?) change during all those months that hh_ide.js was not updated also caused this method to work a bit differently - the occurrences of lvars in function signatures are no longer returned (see two outputs in test plan). Hopefully it doesn't make it much less useful for whatever it does. I'm not going to dig into this now.

Reviewed By: int3

Differential Revision: D2563939

fb-gh-sync-id: d4a05593bce0f32ee99ac0264961857ae1453237