Clean up asm dumps.
Summary: I've fixed up several problems with the existing printing code.  There should no longer be any asm code that isn't printed, the print loop has been reworked so that the assertion about instruction and block ranges is no longer necessary and the dumps for out of order code should be less confusing.

In order to fix out of order assembly, I've used multi-maps to store IRInstruction->TcaRange and Block->TcaRange mappings instead of vectors.  So there can be multiple non-contiguous ranges for each instruction and block.  The main print loop has been rewritten to dump the code for each IRInstruction in machine pc order.  Also a single IRInstruction may be revisited if the assembly code is split by another intervening IRInstruction.

Add validate and dump functions for AsmInfo.  The validate function is used to check that all instruction and block ranges are disjoint.  I check this before and after the relocation phase which has to patch up AsmInfo whenever code is moved.

Reviewed By: ottoni

Differential Revision: D2557502

fb-gh-sync-id: 3d3d3c5aaf63314a7b88e1ccd91801292a27189e